+++ codebase/workers/unity/Assets/Gamelogic/GlobalLayer/NFZ_Templates.cs
@@ -3,7 +3,7 @@ using Improbable.Controller;
 
 public static class NFZ_Templates
 {
-    private static double[] basic = { 
+    private static float[] basic = { 
         25, 25, 
         25, -25,
         -25, -25,
@@ -13,9 +13,9 @@ public static class NFZ_Templates
     public static Improbable.Controller.NoFlyZone GetNoFlyZone(NFZTemplate template)
     {
         Improbable.Controller.NoFlyZone nfz = new Improbable.Controller.NoFlyZone();
-        nfz.vertices = new Improbable.Collections.List<Improbable.Vector3d>();
+        nfz.vertices = new Improbable.Collections.List<Improbable.Vector3f>();
 
-        double[] coords = basic;
+        float[] coords = basic;
         switch (template)
         {
             case NFZTemplate.BASIC:
@@ -25,7 +25,7 @@ public static class NFZ_Templates
 
         for (int i = 0; i < coords.Length; i += 2)
         {
-            nfz.vertices.Add(new Improbable.Vector3d(coords[i], 0, coords[i+1]));
+            nfz.vertices.Add(new Improbable.Vector3f(coords[i], 0, coords[i+1]));
         }
 
         NoFlyZone.setBoundingBoxCoordinates(ref nfz);
